Create Aggregate Functions
Trends in Startups
Howdy! It’s your first day as a TechCrunch reporter. Your first task is to write an article on the rising trends in the startup world.

To get you started with your research, your boss emailed you a project.sqlite file that contains a table called startups. It is a portfolio of some of the biggest names in the industry.

Write queries with aggregate functions to retrieve some interesting insights about these companies.

What are you waiting for? Let’s get started!

If you get stuck during this project or would like to see an experienced developer work through it, click “Get Unstuck“ to see a project walkthrough video.

Tasks
14/14 Complete
Mark the tasks as complete by checking them off
Write the following queries:
1.
Getting started, take a look at the startups table:

SELECT *
FROM startups;
How many columns are there?


Stuck? Get a hint
2.
Calculate the total number of companies in the table.


Stuck? Get a hint
3.
We want to know the total value of all companies in this table.

Calculate this by getting the SUM() of the valuation column.


Stuck? Get a hint
4.
What is the highest amount raised by a startup?

Return the maximum amount of money raised.


Stuck? Get a hint
5.
Edit the query so that it returns the maximum amount of money raised, during ‘Seed’ stage.


Stuck? Get a hint
6.
In what year was the oldest company on the list founded?


Stuck? Get a hint
Let's find out the valuations among different sectors:
7.
Return the average valuation.


Stuck? Get a hint
8.
Return the average valuation, in each category.


Stuck? Get a hint
9.
Return the average valuation, in each category.

Round the averages to two decimal places.


Stuck? Get a hint
10.
Return the average valuation, in each category.

Round the averages to two decimal places.

Lastly, order the list from highest averages to lowest.


Stuck? Get a hint
What are the most competitive markets?
11.
First, return the name of each category with the total number of companies that belong to it.


Stuck? Get a hint
12.
Next, filter the result to only include categories that have more than three companies in them.

What are the most competitive markets?


Stuck? Get a hint
Let's see if there's a difference in startups sizes among different locations:
13.
What is the average size of a startup in each location?


Stuck? Get a hint
14.
What is the average size of a startup in each location, with average sizes above 500?


Stuck? Get a hint
Code Editor